+++ modules/menu/desktop/listView.js
@@ -44,9 +44,6 @@ const ListView = ListViewMain.extend(/** @lends ListView.prototype */{
                 this.render();
             }
         });
-        channel.on({
-            "change:isOutOfRange": this.renderSelectedList
-        }, this);
         this.listenTo(Radio.channel("Map"), {
             "change": function () {
                 this.renderSelectedList();